feat(templates): add HD-2D wuxia template with procedural assets
- templates/hd2d_wuxia: 分层大地图(视差/前景遮挡)、8向角色/4种敌人、特效、中文位图字体图集、UI、BGM/SFX
- tools/: Pillow 程序化出图 + 纯 Python 音效合成 + mapdata.lua 导出(碰撞与画面同源)
- fix(engine): ui_serializer.cpp rapidjson int64 重载在 GCC 下歧义(仅 Linux 构建暴露)
- docs: 远程台式机编译工作流交接文档入库

+++ engine/ecs/ui_serializer.cpp
@@ -478,19 +478,19 @@ void PutVec4(rapidjson::Value& obj, JsonAlloc& alloc, const char* key, const glm
 }
 
 void PutFloat(rapidjson::Value& obj, JsonAlloc& alloc, const char* key, float v) {
-    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(v).Move(), alloc);
+    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(static_cast<int64_t>(v)).Move(), alloc);
 }
 
 void PutInt(rapidjson::Value& obj, JsonAlloc& alloc, const char* key, int v) {
-    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(v).Move(), alloc);
+    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(static_cast<int64_t>(v)).Move(), alloc);
 }
 
 void PutInt64(rapidjson::Value& obj, JsonAlloc& alloc, const char* key, long long v) {
-    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(v).Move(), alloc);
+    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(static_cast<int64_t>(v)).Move(), alloc);
 }
 
 void PutBool(rapidjson::Value& obj, JsonAlloc& alloc, const char* key, bool v) {
-    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(v).Move(), alloc);
+    obj.AddMember(rapidjson::Value(key, alloc), rapidjson::Value(static_cast<int64_t>(v)).Move(), alloc);
 }
 
 void PutString(rapidjson::Value& obj, JsonAlloc& alloc, const char* key, const std::string& v) {
